requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
networkx>=3.2.1
streamlit>=1.44.0
plotly>=5.20.0
praw>=7.7.1
statsmodels>=0.14.0
scipy>=1.12.0
vaderSentiment>=3.3.2
pytrends>=4.9.2
pyvis>=0.3.2
joblib>=1.3.0
tqdm>=4.66.0
